\name{fieldCorr}
\alias{fieldCorr} 
\docType{data}
\title{Set of fields for Report method.}
\description{      
Dataframe with fields potentially used as stratification by 'Report' method, with french and english translation, inserted in output reporting file.
}
\usage{fieldCorr}
\format{
Fields of \emph{fieldCorr} dataframe:
\tabular{rl}{
Field \tab Name of the field as found in COST objects \cr
FR \tab Field translation in french language \cr
EN \tab Complete name of the field \cr
}}


\examples{
data(fieldCorr)
}
\keyword{datasets}
