\name{fieldTitle}
\alias{fieldTitle} 
\docType{data}
\title{Set of titles for Report method.}
\description{      
Dataframe with titles in french and english, called by Report method when creating reporting file.
}
\usage{fieldTitle}
\format{
Fields of \emph{fieldTitle} dataframe:
\tabular{rl}{
Title \tab Numerical code of title \cr
FR \tab Title in french language \cr
EN \tab Title in english language \cr
}}


\examples{
data(fieldTitle)
}
\keyword{datasets}
